Put test_al in one single script
