+++ tests/UNMOCKED_end2end_tests/test_al.py
@@ -6,17 +6,21 @@ import time
 import numpy as np
 
 from lightly.active_learning.agents.agent import ActiveLearningAgent
+from lightly.active_learning.scorers.classification import ScorerClassification
 from lightly.api.api_workflow_client import ApiWorkflowClient
 from lightly.api.bitmask import BitMask
 from lightly.data import LightlyDataset
 from lightly.api.upload import upload_dataset
+from lightly.openapi_generated.swagger_client import SamplingMethod
 
 from lightly.utils import save_embeddings
 
 from lightly.active_learning.config.sampler_config import SamplerConfig
 
 
-def t_est_unmocked_complete_workflow(path_to_dataset: str, token: str, dataset_id: str):
+def t_est_unmocked_complete_workflow(path_to_dataset: str, token: str, dataset_id: str,
+                                     preselected_tag_name: str = None, query_tag_name: str = None,
+                                     with_scores: bool = True):
     # define the api_client and api_workflow
     api_workflow_client = ApiWorkflowClient(host="https://api-dev.lightly.ai", token=token, dataset_id=dataset_id)
 
@@ -45,12 +49,21 @@ def t_est_unmocked_complete_workflow(path_to_dataset: str, token: str, dataset_i
 
     # perform_a_sampling
     print("Starting the AL loop")
-    total_no_samples = api_workflow_client._get_all_tags()[0].tot_size
-    total_currently_chosen_samples = 0
-    agent = ActiveLearningAgent(api_workflow_client, query_tag_name='initial-tag')
-    for batch_size in [1, 2, 5]:
-        sampler_config = SamplerConfig(batch_size=batch_size)
-        chosen_filenames = agent.query(sampler_config=sampler_config)
+
+    agent = ActiveLearningAgent(api_workflow_client,
+                                query_tag_name=query_tag_name, preselected_tag_name=preselected_tag_name)
+    total_currently_chosen_samples = len(agent.labeled_set)
+    total_no_samples = len(agent.unlabeled_set) + total_currently_chosen_samples
+    for iter, batch_size in enumerate([1, 2, 5]):
+        if iter == 0 or not with_scores:
+            sampler_config = SamplerConfig(batch_size=batch_size)
+            chosen_filenames = agent.query(sampler_config=sampler_config)
+        else:
+            predictions = np.random.rand(len(agent.unlabeled_set), 10)
+            predictions_normalized = predictions / np.sum(predictions, axis=1)[:, np.newaxis]
+            al_scorer = ScorerClassification(predictions_normalized)
+            sampler_config = SamplerConfig(batch_size=batch_size, method=SamplingMethod.CORAL)
+            chosen_filenames = agent.query(sampler_config=sampler_config, al_scorer=al_scorer)
         total_currently_chosen_samples += batch_size
         assert (len(chosen_filenames) == total_currently_chosen_samples)
         assert (len(agent.labeled_set) == total_currently_chosen_samples)
@@ -65,12 +78,31 @@ if __name__ == "__main__":
         path_to_dataset = "/Users/malteebnerlightly/Documents/datasets/clothing-dataset-small-master/test"
         token = os.getenv("TOKEN")
         dataset_id = "602e648a42ece4003201adf9"
-    elif len(sys.argv) == 1+3:
-        path_to_dataset, token, dataset_id = (sys.argv[1+i] for i in range(3))
+        query_tag_name = "sharp-images"
+        preselected_tag_name = "preselected_8_images"
+        with_scores = "True"
+    elif len(sys.argv) == 1 + 6:
+        path_to_dataset, token, dataset_id, query_tag_name, preselected_tag_name, with_scores = \
+            (sys.argv[1 + i] for i in range(6))
     else:
-        raise ValueError("ERROR in number of command line arguments."
-                         "Example: python test_al_step3.py this/is/my/dataset TOKEN dataset_id")
-    for i in range(2):
+        raise ValueError("ERROR in number of command line arguments, must be 6."
+                         "Example: python test_al.py this/is/my/dataset "
+                         "TOKEN dataset_id query_tag_name preselected_tag_name True")
+
+    if query_tag_name == "None":
+        query_tag_name = None
+    if preselected_tag_name == "None":
+        preselected_tag_name = None
+    if with_scores == "True":
+        with_scores = True
+    elif with_scores == "False":
+        with_scores = False
+    else:
+        raise ValueError("'with_scores' must be either 'True' or 'False'")
+
+    for i in range(1):
         print(f"ITERATION {i}:")
-        t_est_unmocked_complete_workflow(path_to_dataset, token, dataset_id)
+        t_est_unmocked_complete_workflow(path_to_dataset, token, dataset_id,
+                                         query_tag_name=query_tag_name, preselected_tag_name=preselected_tag_name,
+                                         with_scores=with_scores)
         print("")
